requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-rest-swagger==2.2.0
djangorestframework==3.10.1
drf-yasg==1.16.1
flex==6.14.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonpointer==2.0
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
oauthlib==3.0.2
openapi-codec==1.3.2
packaging==19.0
pprint==0.1
PyJWT==1.7.1
pyparsing==2.4.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
rfc3987==1.3.8
ruamel.yaml==0.15.100
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
splinter==0.11.0
sqlparse==0.3.0
strict-rfc3339==0.7
uritemplate==3.0.0
urllib3==1.25.3
validate-email==1.3
